add GameService, FileService and AddGameServlet

+++ semestr_work/src/main/java/interfaces/RoleRepository.java
@@ -1,5 +1,5 @@
 package interfaces;
 
 public interface RoleRepository {
-    String findByUserId(int userId);
+    String findByUserId(long userId);
 }
